requirements.txt
flask
stripe
authlib
tinydb
python-dotenv